pyproject.toml
[tool.poetry]
name = "wrtn-crawl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.25.0"
docker = "^7.1.0"
pandas = "^2.2.3"
pymysql = "^1.1.1"
pillow = "^11.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
